requirements.txt
streamlit
yfinance
pandas
numpy
google-generativeai
apscheduler
python-dotenv
